requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
python-dotenv==0.19.0
psycopg2-binary==2.9.9
SQLAlchemy==1.4.36
alembic==1.7.7
PyJWT==2.8.0
APScheduler==3.10.1
Werkzeug==2.0.3
click==8.0.4
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
tzlocal==5.0.1
gunicorn==20.1.0
supervisor==4.2.5